menu/requirements.txt
django==2.2.9
djangorestframework==3.11.0
django-filter==2.2.0
Pillow==6.2.1
psycopg2-binary==2.8.4
